+++ samples/list-buffer/ListBuffer.cpp
@@ -1,44 +1,18 @@
 #include "ListBuffer.h"
-#include "metadata/PropertyType.h"
 
 #include <algorithm>
-#include <cstddef>
-#include <cstdint>
+#include <cstring>
 #include <ranges>
 #include <span>
-#include <stddef.h>
 #include <type_traits>
 
+#include "metadata/PropertyType.h"
+
+#include "ListBufferElementView.h"
+#include "ListView.h"
+#include "TypeToListBufferTag.h"
+
 using namespace db;
-
-ListBufferElementView::ListBufferElementView(std::byte* data, size_t size)
-    : _data(data),
-    _size(size)
-{
-}
-
-ListBufferElementView::ListBufferElementView(ListBuffer::iterator begin,
-                                             ListBuffer::iterator end)
-    : _data(&(*begin)),
-    _size(std::distance(begin, end) - _listBufferTagSize)
-{
-}
-
-template <Listable L>
-L ListBufferElementView::getAs() const {
-    static_assert(std::is_trivially_copyable<L>());
-
-    const std::byte* valueStart = _data + _listBufferTagSize;
-
-    L x {};
-    std::memcpy(&x, valueStart, _size);
-    return x;
-}
-
-ListBuffer::ListBufferTag ListBufferElementView::getTag() const {
-    const std::byte& tagByte = *_data;
-    return ListBuffer::ListBufferTag {std::to_integer<uint8_t>(tagByte)};
-}
 
 template <Listable L>
 ListBufferElementView ListBuffer::insert(const L& listItem) {
@@ -89,7 +63,7 @@ ListView ListBuffer::insert(const R& list) {
     }
 
     const size_t sizePrior = _buf.size();
-    const size_t newSize = sizePrior + listSize; 
+    const size_t newSize = sizePrior + listSize;
 
     _buf.resize(newSize);
 
@@ -125,9 +99,5 @@ namespace db {
 template ListBufferElementView ListBuffer::insert<types::Int64::Primitive>(const long&);
 template ListBufferElementView ListBuffer::insert<types::Double::Primitive>(const double&);
 
-template types::Int64::Primitive ListBufferElementView::getAs<types::Int64::Primitive>() const;
-template types::Double::Primitive ListBufferElementView::getAs<types::Double::Primitive>() const;
-
 template ListView ListBuffer::insert(const std::vector<types::Int64::Primitive>&);
-
 }
